requirements.txt
asgiref==3.3.1
automium==0.2.6
automium-web==0.1.1
click==7.1.2
databases==0.4.1
dj-database-url==0.5.0
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
MarkupSafe==1.1.1
pytz==2020.5
SQLAlchemy==1.3.22
sqlparse==0.4.1
Werkzeug==1.0.1
whitenoise==5.2.0
